db-node/requirements.txt
grpcio==1.50.0
protobuf==4.21.9
flask
